pyproject.toml
[tool.poetry]
name = "filedrop"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "*"
pillow = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
